Backend_New/requirements.txt
# FastAPI Backend Requirements
# ============================

# Web Framework
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv

# LangChain & LangGraph
langchain
langchain-openai
langchain-community
langgraph
langgraph-checkpoint

# Database
psycopg2-binary
SQLAlchemy
chromadb  # For semantic query caching

# Utilities
python-multipart
